requirements.txt
gunicorn
dnspython
flask